chatbot_requirements.txt
streamlit==1.35.0
streamlit-extras==0.4.3
googletrans-py==4.0.0
gTTS==2.5.1
torch==2.3.1
langchain==0.2.3
langchain-community==0.2.4
sentence-transformers==3.0.1
PyMuPDF==1.24.5
beautifulsoup4==4.12.3
tabulate==0.9.0
chromadb==0.5.0